Updating QUAD_STRIP to fix anti-aliasing issue for a common situation
If a quad strip is draw, this algorithm will try to draw as few canvas
paths as possible.  If all vertices have the same fill & stroke, only
one path will be drawn - avoiding the anti-aliasing issue.  If the
fill/stroke is not the same for all vertices, then multiple shapes will
be drawn and the anti-aliasing issue will resurface.